Return only relevant records for Fetch
This commit improves the output (response) for the `Fetch` request.

Before this change, the handler would locate and return the whole log
segment containing requested offset. In the new implementation it will
skip over record batches with lower offsets.

Note that it is still possible to return some records with lower offsets
(those in the first output batch). It is up to the client to filter
these records out, but now the count of such records should be very
small.

This change therefore optimizes for the case when the requested offset
is located towards the end of the log file - i.e., reduces network
traffic for consumers that are "up to date" and fetch the most recent
messages.

Known issues & limitations:
 - A very significant drawback of current implementation is that it
   loads and buffers fetched records in memory.
 - Real Kafka uses an additional index file per segment with in-file
   record batch offsets, then during Fetch handling determines the fetch
   location (within the log file), and issues copy from the log file
   directly to the socket.
 - The I/O operation probably issues a `sendfile(2)` syscall or similar
   to send the log file data to the socket without copying them to the
   user-space.