Copy fetch records directly from the log file
This commit changes Fetch record handling so that it does not load all
the records to the memory before serialization. Instead, it copies the
appropriate portion of the log files directly to the output stream.

This implementation uses the `tokio::io::copy`, which does allocate a
small intermediate buffer, but does not load the full file contents into
memory.

Unfortunately, without being able to modify `Cargo.lock`, this cannot be
improved further. That is, we cannot use the `sendfile(2)` syscall,
which would further eliminate the need to copy data into the userspace.

Known issues & limitations:
 - When determining the physical (i.e., withing the log file) position,
   current implementation does a scan of the log file.
 - Real Kafka uses an accompanying `.index` file which contains
   `(base relative offset, physical offset)` pairs, is append-only and
   much smaller than the raw log data.
 - Employing such an (memory-mapped) index  would also remove the need
   to take an exclusive lock to scan the log file (which prevents
   concurrent fetches of the same log)